copy ios and move asset
