+++ packages/js/app/src/views/SignupScreens/Signup1.tsx
@@ -3,7 +3,7 @@ import { Text, View, StyleSheet, TouchableOpacity } from 'react-native';
 import type { createStackNavigator } from '@react-navigation/stack';
 import TextInputBox from '../../components/TextInputBox';
 import colors from '../../constants/Colors';
-import Signup1Background from '../../assets/SignUp1Background.svg';
+import Signup1Background from '../../../assets/SignUp1Background.svg';
 
 type RootStackParamList = {
     Login: undefined;
