+++ packages/js/app/src/views/SignupScreens/Signup2.tsx
@@ -4,7 +4,7 @@ import type { createStackNavigator } from '@react-navigation/stack';
 import Header from '../../components/Header';
 import TextInputBox from '../../components/TextInputBox';
 import colors from '../../constants/Colors';
-import Signup2Background from '../../assets/SignUp2Background.svg';
+import Signup2Background from '../../../assets/SignUp2Background.svg';
 
 type RootStackParamList = {
     Login: undefined;
